+++ kotlin-analysis-api/src/test/kotlin/com/google/devtools/ksp/impl/test/KSPAATest.kt
@@ -422,11 +422,59 @@ class KSPAATest : AbstractKSPAATest() {
         runTest("../test-utils/testData/api/nullableTypes.kt")
     }
 
-    @Disabled
-    @TestMetadata("overridee.kt")
+    @TestMetadata("conflictingOverride.kt")
     @Test
-    fun testOverridee() {
-        runTest("../test-utils/testData/api/overridee.kt")
+    fun testConflictingOverride() {
+        runTest("../test-utils/testData/api/overridee/conflictingOverride.kt")
+    }
+
+    @Disabled
+    @TestMetadata("javaAccessor.kt")
+    @Test
+    fun testJavaAccessor() {
+        runTest("../test-utils/testData/api/overridee/javaAccessor.kt")
+    }
+
+    @TestMetadata("javaAnno.kt")
+    @Test
+    fun testJavaAnno() {
+        runTest("testData/overridee/javaAnno.kt")
+    }
+
+    @TestMetadata("javaOverrideInSource.kt")
+    @Test
+    fun testJavaOverrideInSource() {
+        runTest("../test-utils/testData/api/overridee/javaOverrideInSource.kt")
+    }
+
+    @TestMetadata("noOverride.kt")
+    @Test
+    fun testNoOverride() {
+        runTest("../test-utils/testData/api/overridee/noOverride.kt")
+    }
+
+    @TestMetadata("overrideInLib.kt")
+    @Test
+    fun testOverrideInLib() {
+        runTest("../test-utils/testData/api/overridee/overrideInLib.kt")
+    }
+
+    @TestMetadata("overrideInSource.kt")
+    @Test
+    fun testOverrideInSource() {
+        runTest("../test-utils/testData/api/overridee/overrideInSource.kt")
+    }
+
+    @TestMetadata("overrideOrder.kt")
+    @Test
+    fun testOverrideOrder() {
+        runTest("testData/overridee/overrideOrder.kt")
+    }
+
+    @TestMetadata("primaryConstructorOverride.kt")
+    @Test
+    fun testPrimaryConstructorOverride() {
+        runTest("../test-utils/testData/api/overridee/primaryConstructorOverride.kt")
     }
 
     @TestMetadata("parameterTypes.kt")
